backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pandas
scikit-learn
xgboost
joblib
pyyaml
starlette
